Adjust touchscreen aspect ratio: doing this on the driver level via DT did not work (only swapping axes does), so we use lgvl's option to scale the touchscreen.
The coordinates coming from evdev are 240x320, we resize them to 320x240.

It seems that config data (chipdata) in the touchscreen itself is wrong so data from the chip is not scaled correctly (having 320 for the short edge and 240 for the long).

It also seems that there is no scaling done in the linux input driver stack, the DT info is just passed along, but not used to modify the reported coordinates.